java

+++ trading_framework/src/main/java/org/afob/prices/PriceListener.java
@@ -1,6 +1,5 @@
 package org.afob.prices;
 
-import java.math.BigDecimal;
 
 public interface PriceListener {
 
@@ -9,6 +8,6 @@ public interface PriceListener {
      * @param productId the id of the product for which the price is given
      * @param price current price of the product
      */
-    void priceTick(String productId, BigDecimal price);
+    void priceTick(String productId, double price);
 
 }
